requirements.txt
fastapi
gunicorn
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pandas 
tweety-ns 
textblob 
alpaca-py 
